logic: vreg_analysis: fix `type` rule
Theoretically, a result type of GEP instruction is a pointer to
the type of addressed element, but sometimes it is convenient to
treat it as  just the addressed type. Thus let's use both types for now.
Example:
  %8 = getelementptr [7 x [6 x i8], [7 x [6 x i8]]* %7, i64 0, i64 5
  %9 = getelementptr [7 x i8], [6 x i8]* %8, i64 0, i64 0
here %9 will have 'i8 *' (correct) and 'i8' (not correct) types,
and %8 - '[6 x i8] *' (correct) and '[6 x i8]' (not correct) types, but
%8 is convenient to consider as [6 x i8] type, especially in `buff_size`

TODO: come up with a better solution for GEP cases