parser: Initial implementation for data flow
The initial implementation of the data flow graph inside a function.
In this implementation, a sink is always an instruction, while a source
can take on one of four forms: an instruction, a constant, a global variable,
or a function argument.

The new fact is added `data_flow_local` with format:
  `<id>;<func id>;<block id>;<instr id>;<src type>;<source id>`

`<src type>` corresponds to one of four values and affects how `<source id>`
should be interpreted::
0 - instruction, `<source id>` is `instr_t`
1 - constant, `<source id>` is `const_t`
2 - global variable, `<source id>` is `global_t`
3 - function argument, `<source id>` is `func_arg_t`.