generator: logic: Update CFG facts
Originally, CFG facts had function and block names, not identifiers.
This made them a bit difficult to use, since a conversion to an
identifier was required. This was done because the llvm-ir-analysis
crate uses names.
Fixed 'generator' to do the conversion itself, which should improve
overall performance in linked rules.