logic: add `DataFlowGraph.sources_vreg`
Add `DataFlowGraph.sources_vreg` relation to get all VREG source for
given one.

It would be helpful to include an instruction ID along with the VREG ID
when returning a source, since the instruction ID can be used to narrow
down other relations that use sources_vreg. However, there is a challenge
with doing so for function arguments or global variables since there is no
instruction ID associated with them.

A possible workaround is to return the first instruction ID of the function,
but this is not a perfect solution.